made it double and propagate exception to view

+++ Assignment - Operazioni Matematiche/src/model/Moltiplicazione.java
@@ -7,7 +7,7 @@ public class Moltiplicazione extends Operation{
 	}
 
 	@Override
-	protected int doOperation() {
+	protected double doOperation() {
 		return this.a*this.b;
 	}
 }
